Add ability to overwrite the ensemble background files
Previous code version crashed when in_path and out_path were the same in
the input YAML file. Updated code checks whether in_path and out_path
are identical, and then opens the netCDF file in "append" mode if that
is the case, which allows the user to overwrite the ensemble background
files.

A new unit test was added to tests/test_ens_io.py to verify that this
feature works. All pytest unit tests have passed.