Separate computation of horizontal and vertical localization
Original code created a massive array with all (x, y, z) coordinates within the
model domain, and used that to compute localization. This is rather inefficient.
This run time for this method is proportional to N2d X Nz, where N2d is the
number of points in the horizontal mesh amd Nz is the number of vertical levels.

The new code computes the localization for the horizontal and vertical directions
separately, and should have run time proportional to N2d + Nz.

All pytests pass with the updated code and the test case runs w/out error.
The resulting C localization array from the new method does differ from
the old method, but the maximum differences are O(1e-5), and the O-A values
are the same out to 4 decimal points in the test case, so these differences
should be small enough to be insignificant.